% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pvalues.R
\name{compute_mu}
\alias{compute_mu}
\title{Compute Mu for MZINB Model}
\usage{
compute_mu(data, X, eta, beta)
}
\arguments{
\item{data}{The count data matrix.}

\item{X}{The covariates matrix.}

\item{eta}{The \code{eta} vector.}

\item{beta}{The \code{beta} matrix.}
}
\value{
The computed \code{mu} matrix.
}
\description{
This function computes the \code{mu} matrix based on the current values of \code{eta} and \code{beta}.
}
