% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pvalues.R
\name{compute_pvalues}
\alias{compute_pvalues}
\title{Compute P-values for MZINB Model}
\usage{
compute_pvalues(params)
}
\arguments{
\item{params}{A list containing final estimates of model parameters.}
}
\value{
A list containing the computed p-values for each feature.
}
\description{
This function calculates the p-values for the regression coefficients \code{beta} in the MZINB model
based on the computed Fisher information matrix.
}
