% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM-algorithm.R
\name{e_step}
\alias{e_step}
\title{Perform the E-Step of the EM Algorithm}
\usage{
e_step(mu, tau, data, X, eta, beta, pi, theta, v)
}
\arguments{
\item{mu}{A numeric matrix (n x m) representing the mean values of the count data. This matrix will be updated during the E-step.}

\item{tau}{A numeric matrix (n x m) representing the posterior probabilities for zero inflation. This matrix will be updated during the E-step.}

\item{data}{A numeric matrix (n x m) of observed count data, where rows represent samples and columns represent features (e.g., genes or taxa).}

\item{X}{A numeric matrix (n x p) of covariates for the samples, where rows correspond to samples and columns represent covariates.}

\item{eta}{A numeric vector (length n) of sample-specific offsets.}

\item{beta}{A numeric matrix (m x p) of covariate effects for each feature.}

\item{pi}{A numeric vector (length m) of zero-inflation probabilities for each feature.}

\item{theta}{A numeric vector (length m) of dispersion parameters for each feature.}

\item{v}{A binary matrix (n x m) indicating zero counts (1 if zero, 0 otherwise).}
}
\value{
A list containing the updated values for \code{mu} and \code{tau}:
\describe{
\item{mu}{A numeric matrix (n x m) of updated mean values for the count data.}
\item{tau}{A numeric matrix (n x m) of updated posterior probabilities for zero inflation.}
}
}
\description{
The \code{e_step} function performs the Expectation step (E-step) of the Expectation-Maximization (EM) algorithm for the Multivariate Zero-Inflated Negative Binomial (MZINB) model.
In this step, the function computes the expected values of the latent variables (\code{mu} and \code{tau}) based on the current parameter estimates.
}
