% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_mzinb.R
\name{initialize_mzinb}
\alias{initialize_mzinb}
\title{Initialize Parameters for the MZINB Model}
\usage{
initialize_mzinb(
  counts,
  covariates,
  lambda = 0.5,
  beta_bound = c(-10, 10),
  theta_bound = c(0.01, 10),
  eta_bound = 10
)
}
\arguments{
\item{counts}{A numeric matrix (n x m) of count data, where rows represent samples and columns represent features.}

\item{covariates}{A numeric matrix (n x p) of covariates for the samples, where rows correspond to samples and columns correspond to covariates.}

\item{lambda}{A regularization parameter for controlling model complexity. Default is \code{0.5}.}

\item{beta_bound}{A numeric vector specifying the lower and upper bounds for the \code{beta} parameter. Default is \code{c(-10, 10)}.}

\item{theta_bound}{A numeric vector specifying the lower and upper bounds for the \code{theta} parameter. Default is \code{c(1e-2, 10)}.}

\item{eta_bound}{A numeric value specifying the upper bound for the \code{eta} parameter. Default is \code{10}.}
}
\value{
A list containing initialized parameters:
\itemize{
\item \code{data}: Count data matrix.
\item \code{X}: Covariates matrix.
\item \code{n}: Number of samples.
\item \code{m}: Number of features.
\item \code{p}: Number of covariates.
\item \code{pi}: Initial zero-inflation probabilities.
\item \code{eta}: Initial \code{eta} values computed from the non-zero counts.
\item \code{eta_bound}: Updated upper bound for \code{eta}.
\item \code{beta}: Initial regression coefficients.
\item \code{mu}: Initial mean values for the count data.
\item \code{theta}: Initial dispersion parameters.
\item \code{v}: Binary matrix indicating zero counts.
\item \code{tau}: Matrix of initial posterior probabilities for zero inflation.
\item \code{lambda}: Regularization parameter.
\item \code{beta_bound}: Bounds for the \code{beta} parameter.
\item \code{theta_bound}: Bounds for the \code{theta} parameter.
}
}
\description{
The \code{initialize_mzinb} function sets up the initial parameters for the Multivariate Zero-Inflated Negative Binomial (MZINB) model.
It converts the count and covariate data into matrices and initializes values for \code{pi}, \code{eta}, \code{beta}, \code{mu}, \code{theta}, \code{v}, and \code{tau}.
}
\examples{
\donttest{
# Example usage
counts <- matrix(rpois(100, lambda = 10), nrow = 10)
covariates <- matrix(runif(20), nrow = 10)
params <- initialize_mzinb(counts, covariates)
}

}
