% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM-algorithm.R
\name{m_step}
\alias{m_step}
\title{M-Step for MZINB Model}
\usage{
m_step(params)
}
\arguments{
\item{params}{A list containing current estimates of model parameters.}
}
\value{
A list containing updated values for \code{pi}, \code{theta}, and \code{beta}.
}
\description{
This function implements the M-step for the MZINB model, optimizing \code{pi}, \code{theta}, and \code{beta}
using the current estimates of latent variables.
}
