% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_eta.R
\name{update_eta}
\alias{update_eta}
\title{Update Eta for MZINB Model}
\usage{
update_eta(params)
}
\arguments{
\item{params}{A list containing current estimates of model parameters.}
}
\value{
A list containing updated values for \code{eta}.
}
\description{
This function updates the \code{eta} parameter for each observation during the M-step of the MZINB model.
}
