% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM-algorithm.R
\name{e_step}
\alias{e_step}
\title{E-Step for MZINB Model}
\usage{
e_step(params)
}
\arguments{
\item{params}{A list containing current estimates of model parameters (as initialized by \code{initialize_parameters()}).}
}
\value{
A list containing updated values for \code{tau}, the latent variable matrix, and \code{mu}, the expected counts.
}
\description{
This function implements the E-step for the MZINB model, calculating the latent variable \code{tau}
based on current estimates of the model parameters.
}
