% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM-algorithm.R
\name{update_parameters}
\alias{update_parameters}
\title{Update Parameters During the M-Step (Rcpp Version)}
\usage{
update_parameters(
  data,
  X,
  eta,
  beta,
  theta,
  pi,
  tau,
  v,
  lambda,
  beta_bound,
  theta_bound
)
}
\arguments{
\item{data}{A numeric matrix (n x m) of observed count data, where rows represent samples and columns represent features.}

\item{X}{A numeric matrix (n x p) of covariates for the samples.}

\item{eta}{A numeric vector (n) of sample-specific offsets.}

\item{beta}{A numeric matrix (m x p) of covariate effects for each feature.}

\item{theta}{A numeric vector (m) of dispersion parameters for each feature.}

\item{pi}{A numeric vector (m) of zero-inflation probabilities for each feature.}

\item{tau}{A numeric matrix (n x m) of posterior probabilities for zero inflation.}

\item{v}{A binary matrix (n x m) indicating zero counts (1 if zero, 0 otherwise).}

\item{lambda}{A regularization parameter for controlling model complexity.}

\item{beta_bound}{A numeric vector specifying the bounds for the \code{beta} parameter.}

\item{theta_bound}{A numeric vector specifying the bounds for the \code{theta} parameter.}
}
\value{
A list containing updated values for \code{pi}, \code{theta}, and \code{beta}.
}
\description{
This function updates the \code{pi}, \code{theta}, and \code{beta} parameters during the Maximization (M) step of the EM algorithm.
It uses Rcpp for improved performance and efficiency.
}
\details{
This function is part of the M-step of the EM algorithm for the Multivariate Zero-Inflated Negative Binomial (MZINB) model.
It optimizes \code{theta} and \code{beta} using C++ for improved speed and performance.
}
